Revert adding wrong check
Since https://github.com/WordPress/gutenberg/pull/22520 was merged
on May we've been enqueuing the presets as CSS variables
whether or not the theme had support for theme.json
as a way to speed up themes using them and get feedback.

There are enough checks in place that this is inocuous.